LayoutTests/media/track/track-webvtt-tc018-align-text-line-position-expected.txt

Tests cue alignment, line and text position from settings.


*** Testing text track 0
EXPECTED (cues.length == '5') OK
EXPECTED (cues[0].align == 'start') OK
EXPECTED (cues[0].position == '10') OK
EXPECTED (cues[0].line == '0') OK
EXPECTED (cues[0].snapToLines == 'false') OK
EXPECTED (cues[1].align == 'start') OK
EXPECTED (cues[1].position == '50') OK
EXPECTED (cues[1].line == '0') OK
EXPECTED (cues[1].snapToLines == 'true') OK
EXPECTED (cues[2].align == 'middle') OK
EXPECTED (cues[2].position == '80') OK
EXPECTED (cues[2].line == '80') OK
EXPECTED (cues[2].snapToLines == 'false') OK
EXPECTED (cues[3].align == 'end') OK
EXPECTED (cues[3].position == '30') OK
EXPECTED (cues[3].line == '5') OK
EXPECTED (cues[3].snapToLines == 'true') OK
EXPECTED (cues[4].align == 'middle') OK
EXPECTED (cues[4].position == '60') OK
EXPECTED (cues[4].line == '-3') OK
EXPECTED (cues[4].snapToLines == 'true') OK

*** Testing text track 1
EXPECTED (cues.length == '3') OK
EXPECTED (cues[0].align == 'middle') OK
EXPECTED (cues[0].position == '50') OK
EXPECTED (cues[0].line == '-1') OK
EXPECTED (cues[0].snapToLines == 'true') OK
EXPECTED (cues[1].align == 'end') OK
EXPECTED (cues[1].position == '0') OK
EXPECTED (cues[1].line == '-1') OK
EXPECTED (cues[1].snapToLines == 'true') OK
EXPECTED (cues[2].align == 'middle') OK
EXPECTED (cues[2].position == '60') OK
EXPECTED (cues[2].line == '-3') OK
EXPECTED (cues[2].snapToLines == 'true') OK
END OF TEST
